[357386] Move Bundle and Console server editor parts from Eclipse Virgo
The Bundle Information, Bundle Dependency Graph and Server Console
server editor parts from Virgo IDE are now moved to Eclipse Libra. 

Along with the move several changes are done:
1. Removed are dependencies to SpringFramework and Mylyn plugins. 
2. Removed is the tight coupling to the Virgo Manifest editor. The "open
bundle manifest" action is re-implemented in more generic way.
It checks for available IContentType and finds the editor that
corresponds to it. The action first try to find if there any special
content types registered for MANIFEST.MF (like the Virgo Manifest
editor). If none are found, then it tries to retrieve the content type
for the PDE Manifest. If PDE is not present, then it falls back to the
default plain text content type. 
3. The editor is decoupled from the Virgo Server model: communication
from between the editor parts and server runtime is done via interfaces
which the server runtime must implement. 
4. Namespace of the Java classes is changed to org.eclipse.libra.*

The code is organized in two bundles:
- org.eclipse.libra.framework.editor.core - contains all interfaces used
by the editor parts that has to be implemented by the server rutime. 
- org.eclipse.libra.framework.editor.ui - contains the code of the
editor parts.